requirements.txt
crewai==0.86.0
crewai-tools==0.12.1
langchain-groq==0.2.1
python-dotenv==1.0.0
gitpython==3.1.43
